move device logic into adapter
