+++ example_commands.py
@@ -1,21 +1,9 @@
+from divoom_adapter import DivoomAdapter
 import divoom_image
-import divoom_protocol
-import divoom_device
 import time
 from PIL import Image
 
-class DivoomExamples:
-
-    def __init__(self, addr):
-        self.protocol = divoom_protocol.DivoomAuraBoxProtocol()
-        self.device = divoom_device.DivoomDevice(addr)
-
-    def __enter__(self):
-        self.device.connect()
-        return self
-
-    def __exit__(self, exc_type, exc_value, traceback):
-        self.device.disconnect()
+class ExampleCommands(DivoomAdapter):
 
     def show_files(self, filelist, delay=1):
         for f in filelist:
